+++ CHANGES.rst
@@ -10,6 +10,7 @@ in progress
 - Caching: Optionally configure TTL using environment variable ``CACHE_TTL``
 - History: Stop ``grafana-wtf log <UID>`` acquiring *all* dashboards
 - Refactoring: Move all report renderers to ``grafana_wtf.report``
+- History: Add ``id`` and ``uid`` dashboard attributes to report
 
 2023-03-05 0.14.1
 =================